feat: Add Slider component and update Header
This commit adds the Slider component to the client application and updates the Header component to include the Slider and a secondary call-to-action button. The Slider component will be responsible for displaying a slideshow of images. This change enhances the user experience by providing more interactive content on the header section of the page.